backend/requirements.txt
fastapi
uvicorn
mlflow
python-jose
passlib[bcrypt]
python-dotenv
openai
